refact: move Dto-s to packages
