+++ src/main/java/com/healingsys/dto/department/SimpleDepartmentDetailsDto.java
@@ -1,7 +1,6 @@
-package com.healingsys.dto;
+package com.healingsys.dto.department;
 
 import com.healingsys.entities.DayOfWeek;
-import com.healingsys.entities.enums.Status;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,13 +11,10 @@ import java.util.Set;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class DepartmentDetailsDto {
+public class SimpleDepartmentDetailsDto {
+    private Long id;
     private String name;
     private LocalTime opening;
     private LocalTime closing;
-    private int maxGeneratedDays;
-    private double slotLengthInHour;
-    private int slotMaxCapacity;
     private Set<DayOfWeek> closedDay;
-    private Status status;
 }
